requirements.txt

numpy
scipy
matplotlib
pyyaml
